first pass at shared model support
This splits out the variable info from the diagram info
Each node has a corresponding variable reference.
The diagram views do not use the variable reference directly.
There will be times when the variable reference is not available yet, or has been deleted.
So by proxying the variable through the node, the node can take care of these
cases when the variable is not available.

A few ugly parts:
- in the react-flow diagram, the ids used are the variable ids instead of the node ids.
  This is because the variables maintain the links between the nodes. So to generate
  the tree for react-flow we'd have to look up the node for each related variable. And doing
  that look up required a circular reference where the node is asking its parent for something.
  Probably this can be improved.
- the DQRoot has a reference to the Variables model. When this is used in CLUE, I was thinking
  of replacing the Variables model with an extended version that had any necessary shared model
  stuff. Plus this reference would then be stored both in the DQRoot and in the ToolTileModel
  wrapping the DQRoot. This can be replaced with an environment object passed to the DQRoot
  instead. Then only one reference needs to be stored.
- the storage format has now changed, so any old CODAP docs or shared links will be broken.