refactor so shared variables work when inside of CLUE
The main change was switching the variables reference on DQRoot to be a volatile property.
And now it is an API instead of a directly MST model.

It makes more sense as a environment property that can be set with dependency injection,
however that approach is complicated because in CLUE this means we'd have to set this
environment property up at the documents store level. A volatile property lets us scope this
value to just this specific content model.

The API is used instead of the content model so the shared variables implementation in CLUE
can replace the simple implementation here.