fix: Edit Variable dialog displays read only expressions (PT-183974159)
PT Story: https://www.pivotaltracker.com/story/show/183974159

This is a fix for a follow up comment/request in the story about not allowing line breaks in the variable edit dialog's Value and Unit fields. Also included is a small style change to prevent scrollbars from appearing in textareas that do not yet have scrollable content.