fix: Long values in variable dialog not displayed properly on card (PT-184474236)
PT Story: https://www.pivotaltracker.com/story/show/184474236

This change makes the ExpandableInput component properly update its styling for long values after the value is set using the Edit Variable dialog. Previously, if a long value was set in the Edit Variable dialog, the input field was not changing to indicate that the value was longer than the field.