fix: Undefined in expression calculation (PT-184955306)
PT Story: https://www.pivotaltracker.com/story/show/184955306

Add a check to `calculationString` to make sure each node has an associated input. If not, simply return the node instead of attempting to return a non-existent value and unit. This fixes a bug where, for example, `calculationString` would convert an expression like `a+b in ft` to `10 cm + 10 mm in undefined undefined`. It will now correctly convert it to `10 cm + 10 mm in ft`.